+++ src/test/java/web/steps/NavigationSteps.java
@@ -4,37 +4,30 @@ import eu.ibagroup.junase.model.util.Assert;
 import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
-import web.screens.NavigationScreen;
+import web.screens.Navigation;
 
 public class NavigationSteps {
 
-    private final NavigationScreen navigation = new NavigationScreen();
+    private final Navigation navigation = new Navigation();
 
-    @Given("^I am on screen (.*) in application (Sap)$")
-    @Then("^I assert I am on screen (.*) in application (Sap)$")
+    private static final String DELIMITER = " -> ";
+
+    @Given("^I am on screen (.*) in application (Onliner)$")
+    @Then("^I assert I am on screen (.*) in application (Onliner)$")
     public void assertScreenInApplication(String screen, String app) {
         Assert.assertTrue(() -> navigation.getScreenTitle().contains(screen));
     }
 
-    @When("^I set (.*) in field (.*) on screen (.*) in application (Sap)$")
-    public void setValueInField(String value, String iconName, String screen, String app) {
-        navigation.clickIconByName(iconName);
-        navigation.setValue(value, iconName);
+    @When("^I navigate ((?!.*search).*) on screen (Onliner) in application (Onliner)$")
+    public void navigateSection(String section, String screen, String app) {
+        navigation.navigateSection(section);
     }
 
-    @When("^I switch to (.*) on screen (.*) in application (Sap)$")
-    public void switchToTab(String tabName, String screen, String app) {
-        navigation.clickTabByName(tabName);
-    }
-
-    @Then("^I assert modal window (.*) is opened on screen (.*) in application (Sap)$")
-    public void assertModalWindowIsDisplayed(String modalName, String screen, String app) {
-        navigation.switсhIFrameSelectTransactionType();
-        Assert.assertEquals(modalName, navigation::getModalWindowText);
-    }
-
-    @When("^I navigate to type (.*) on modal window (.*) on screen (.*) in application (Sap)$")
-    public void navigateTypeOnModalWindow(String type,String modalName, String screen, String app) {
-        navigation.navigateType(type);
+    @When("^I navigate (.*) on screen (Kaталог Onliner) in application (Onliner)$")
+    public void navigateTabs(String tabs, String screen, String app) {
+        String[] menuItems = tabs.split(DELIMITER);
+        navigation.navigateToFirstLevel(menuItems[0]);
+        navigation.navigateToSecondLevel(menuItems[1]);
+        navigation.navigateToThirdLevel(menuItems[2]);
     }
 }
